requirements-macos.txt
# --------------------------
# Python version
# Requires Python 3.9
# pip install will ignore is version mismatches
python == "3.9"
# --------------------------
# Packages for macOS
pytorch
librosa
tqdm
numpy == "1.24.3"